# docs/openapi.yml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Symfony Demo
  description: Symfony Demo
paths:
  /api/club:
    put:
      summary: Create club
      operationId: CreateClub
      tags:
        - Club
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Club'
      responses:
        '201':
          description: Create club
        '400':
          description: Name or budget is not valid
...
  /api/player:
    put:
      summary: Create Player
      operationId: CreatePlayer
      tags:
        - Player
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Player'
      responses:
        '201':
          description: Player created
        '400':
          description: Name is not valid
        '404':
          description: Club not found
        '409':
          description: Insufficient budget to add this player to the club
    get:
      summary: Get Players
      operationId: GetPlayers
      tags:
        - Player
      parameters:
        - name: orderBy
          in: query
          description: parameter to order by
          required: false
          schema:
            type: string
            default: 'desc'
        - name: club
          in: query
          description: filter by club
          required: false
          schema:
            type: string
            default: null
        - name: sortBy
          in: query
          description: field to sort by (name, salary)
          required: false
          schema:
            type: string
            default: 'salary'
        - name: limit
          in: query
          description: max number of players
          required: false
          schema:
            type: integer
            default: 50
        - name: page
          in: query
          description: page number
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Get Player list
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    example: 'success'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Player'
...
  /api/player/{name}/club:
    delete:
      summary: Remove player's club
      operationId: RemovePlayerClub
      tags:
        - Player
      parameters:
        - name: name
          in: path
          description: player's name
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Club removed from player
...
components:
  schemas:
    Club:
      properties:
        name:
          type: string
          required: true
          maximum: 255
        club:
          type: string
          required: false
        budget:
          type: number
          required: true
          minimum: 0
...
    Player:
      properties:
        name:
          type: string
          required: true
        club:
          type: string
          required: false
          nullable: true
        salary:
          type: number
          required: true
          minimum: 0